Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
sendgrid==6.11.0